Refactor RNG cache utilities into shared module
